requirements.txt
aiocron==1.8
aiofiles==23.2.1
aiogram==2.14.3
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
Babel==2.14.0
certifi==2023.11.17
croniter==2.0.1
frozenlist==1.4.1
idna==3.6
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.9.0
tzlocal==5.2
yarl==1.9.4
